i2c: Convert return type of i2c_send_data to uint8_t
We don't need 16 bits for the return value of i2c_send_data.  Convert
it to uint8_t.

+++ src/i2c-gpio.c
@@ -89,7 +89,7 @@ static uint8_t receive_bit (int m) {
         return r;
 }
 
-int i2c_send_data (int master, uint8_t data) {
+uint8_t i2c_send_data (int master, uint8_t data) {
         send_bit(master, (data >> 7));
         send_bit(master, (data >> 6));
         send_bit(master, (data >> 5));
