+++ src/i2c.h
@@ -14,5 +14,5 @@
 void i2c_get (int m);
 void i2c_send_start (int m);
 void i2c_send_stop (int m);
-int i2c_send_data (int master, uint8_t data);
+uint8_t i2c_send_data (int master, uint8_t data);
 uint8_t i2c_receive_data (int master);
